Make decrypt method read api_key and api_secret from secrets.json
The decrypt method on the Bittrex class was not actually reading
credentials from secrets.json. This commit adds that.

Justification:
I don't currently understand how the decrypt method is supposed
to be used. The way I understand this and would personally use
python-bittrex in production is as follows:

1. I would first create a secrets.json file (using for example
the static encrypt function or preferrably a dedicated secrets.json
creator). This could be done anywhere, for example offline.
2. I would then upload the secrets file to an Internet-connected
machine such as a VPS server and start running a trading script
that utilizes python-bittrex as a dependency.
3. When starting up the trading script, the trading script utilizing
python-bittrex would call the decrypt method, I would provide the password
for decripting credentials from secrets.json and decrypt method then
carries on and the trading script can continue doing its thing.
4. Now if the machine that the trading script was on gets compromized,
the api_key and api_secret are not leaked.

In summary: this commit aims to make the above a reality.

Suggested further improvements: standalone secrets.json generator